requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
datasets==3.6.0
dill==0.3.8
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
pyarrow==20.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==78.1.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.15
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
